backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.8
sqlalchemy==1.4.41
pydantic==1.10.13
python-dotenv==1.0.0
itsdangerous==2.1.2
python-multipart==0.0.6
